pyproject.toml
[project]
name = "riceshare"
version = "0.1.0"
authors = [
    { name="Nub-programmer", email="nub@example.com" }
]
description = "CLI tool for discovering and installing Linux rice profiles"
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "typer>=0.9.0",
    "requests>=2.31.0",
    "pyyaml>=6.0.0",
    "rich>=13.7.0"
]
[project.scripts]
riceshare = "riceshare.cli:app"
